Remove anime.interface.ts
