+++ apps/angular/src/core/services/mappers/anime-list-options.mapper.ts
@@ -2,7 +2,6 @@ import { HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { isSortField, isType } from '@js-camp/core/utils/guards/sort.guard';
-import { AnimeListOptions } from '@js-camp/core/utils/interfaces/anime.interface';
 import { SortDirection, SortField } from '@js-camp/core/utils/types/sort';
 
 /** All possible query parameters. */
@@ -20,6 +19,42 @@ namespace DefaultParamValue {
   export const SEARCH = '';
   export const ORDERING = `${SortDirection.Ascending}${SortField.TitleEnglish}`;
   export const TYPE = '';
+}
+
+/** Sort setting for anime list request. */
+interface SortSetting {
+
+  /** The field by which to sort. */
+  field: string;
+
+  /** The sort direction. */
+  direction: string;
+}
+
+/** Filter setting for anime list request. */
+interface FilterSetting {
+
+  /** All possibly types. */
+  byType: string[];
+}
+
+/** Params for for anime list request. */
+export interface AnimeListOptions {
+
+  /** The page number to be returned. */
+  readonly pageNumber: number;
+
+  /** Sort setting. */
+  readonly sort: SortSetting;
+
+  /** Filter setting. */
+  readonly filter: FilterSetting;
+
+  /** Search query. */
+  readonly search: string;
+
+  /** Maximum number of entries per page.*/
+  readonly limit?: number;
 }
 
 /** Params mapper. */
